*: move prepare checker to coordinator (#4594)
* remove cluster from coordinator


* move prepare checker to coordinator
